Debug flight Model

+++ gym_environment.py
@@ -1,5 +1,6 @@
 import os
 import gym
+import logging
 import pandas as pd
 from gym import spaces
 import numpy as np
@@ -9,10 +10,11 @@ import matplotlib.pyplot as plt
 import pyglet
 import shutil
 from numpy.linalg import norm
+from utils import setup_logger
+from configparser import ConfigParser
 
 TAKE_OFF_ALTITUDE = converter(80, "feet", "m")  # 80 feets
 RUNWAY_LENGTH = 5000  # 5000m
-from configparser import ConfigParser
 
 
 parser = ConfigParser()
@@ -23,10 +25,25 @@ parser.read(config_path)
 DELTA_T = eval(parser.get("flight_model", "Timestep_size"))
 LEVEL_TARGET = converter(eval(parser.get("task", "LEVEL_TARGET")), "feet", "m")
 MAX_TIMESTEP = 200 / DELTA_T
+DEBUG = eval(parser.get("debug", "debug"))
+
+# create and configure logger
+os.makedirs("logs", exist_ok=True)
+if DEBUG:
+    level = logging.DEBUG
+else:
+    level = logging.INFO
+LOG_FORMAT = "%(levelno)s %(asctime)s %(funcName)s - %(message)s"
+logger = setup_logger(
+    "gym_logger",
+    "logs/gym.log",
+    level=level,
+    format=LOG_FORMAT,
+)
 
 
 class PlaneEnv(gym.Env):
-    metadata = {"render.modes": ["human", "rgb_array"], "video.frames_per_second": 50}
+    metadata = {"render.modes": ["human", "rgb_array"], "video.frames_per_second": 60}
     """Custom Environment that follows gym interface"""
 
     def __init__(self, task="take-off"):
@@ -56,7 +73,9 @@ class PlaneEnv(gym.Env):
         self.rewards = []
         self.rewards_1 = []
         self.rewards_2 = []
-        shutil.rmtree("trajectories")
+        self.distance_vec = []
+        if "trajectories" in os.listdir():
+            shutil.rmtree("trajectories")
         os.makedirs("trajectories", exist_ok=True)
 
     def terminal(self):
@@ -84,8 +103,9 @@ class PlaneEnv(gym.Env):
         elif self.task == "level-flight":
             self.overtime = self.FlightModel.timestep > MAX_TIMESTEP
             self.overspeed = self.FlightModel.Mach > 0.98
-            self.over_g = norm(self.FlightModel.A) > (1.5 * 9.81)
-
+            self.g = norm(self.FlightModel.A) / 9.81
+            self.over_g = self.g > 1.5
+            self.out = abs(self.FlightModel.Pos[1] - LEVEL_TARGET) > 1000
             if self.overtime:
                 self.reason_terminal = "Overtime"
                 return True
@@ -96,6 +116,9 @@ class PlaneEnv(gym.Env):
                 self.reason_terminal = "Over G"
                 # print(norm(self.FlightModel.A))
                 return False
+            if self.out:
+                self.reason_terminal = "out"
+                return True
             else:
                 return False
 
@@ -109,15 +132,20 @@ class PlaneEnv(gym.Env):
             if self.take_off:
                 reward = 3000 - 2 * (self.FlightModel.Pos[0] / (RUNWAY_LENGTH / 10))
         elif self.task == "level-flight":
-            reward = (
-                20
-                - abs(LEVEL_TARGET - self.FlightModel.Pos[1])
-                - norm(self.FlightModel.A)
-            )   
-            if self.FlightModel.Mach > 0.90:
-                reward += -100
-            if self.over_g:
+            distance = abs(LEVEL_TARGET - self.FlightModel.Pos[1])
+            self.distance_vec.append(distance)
+            reward = -distance / 1000  # * max(self.g, 1)
+            # 20
+
+            #     - norm(self.FlightModel.A)
+            # )
+            # if self.FlightModel.Mach > 0.90:
+            #     reward += -100
+            if self.out:
                 reward += -1000
+
+            # if self.over_g or self.out:
+            #     reward += -1000
 
         return reward
 
@@ -127,14 +155,23 @@ class PlaneEnv(gym.Env):
         done = self.terminal()
         reward = self.compute_reward(obs)
         self.rewards.append(reward)
+        if done:
+            logger.info(
+                f"Terminal : reason  {self.reason_terminal}, sum of rewards {sum(self.rewards)}"
+            )
         # self.rewards_1.append(reward_1)
         # self.rewards_2.append(reward_2)
         if done:
             self.sum_rewards = np.sum(self.rewards)
-            if self.episode % 1000 == 0:
-                print(
-                    f"Episode {self.episode}, State : {[np.round(x,2) for x in obs]}, Sum of rewards {np.round(self.sum_rewards,0)}, Episode length {self.FlightModel.timestep}, Result {self.reason_terminal}"
-                )
+            if self.episode % 50 == 0:
+                if self.task == "take-off":
+                    print(
+                        f"Episode {self.episode}, State : {[np.round(x,2) for x in obs]}, Sum of rewards {np.round(self.sum_rewards,0)}, Episode length {self.FlightModel.timestep}, Result {self.reason_terminal}"
+                    )
+                elif self.task == "level-flight":
+                    print(
+                        f"Episode {self.episode}, sum of distances to the target {np.round(np.sum(self.distance_vec),0)},Sum of rewards {np.round(self.sum_rewards,0)}"
+                    )
                 fig, ax = plt.subplots()
                 plt.plot(self.FlightModel.Pos_vec[0], self.FlightModel.Pos_vec[1])
                 plt.title("Trajectory")
@@ -216,6 +253,7 @@ class PlaneEnv(gym.Env):
         self.rewards = []
         self.rewards_1 = []
         self.rewards_2 = []
+        self.distance_vec = []
         return np.array(self.FlightModel.obs)
 
     def render(self, mode="human"):
@@ -252,8 +290,9 @@ class PlaneEnv(gym.Env):
             )
 
             axleoffset = cartheight / 4.0
-            cart = rendering.FilledPolygon([(l, b), (l, t), (r, t), (r, b)])
-            # cart = rendering.Image("plane.png", 1.0, 1.0)
+            # cart = rendering.FilledPolygon([(l, b), (l, t), (r, t), (r, b)])
+            cart = rendering.Image("A320_R.png", 300, 100)
+            # cart = rendering.Image()
 
             self.carttrans = rendering.Transform()
             cart.add_attr(self.carttrans)
